# Introduce support for styling documents with ANSI codes