--- a/doclayout.cabal
+++ b/doclayout.cabal
@@ -24,6 +24,7 @@ cabal-version:       >=1.10
 library
   hs-source-dirs:      src
   exposed-modules:     Text.DocLayout
+  other-modules:       Text.DocLayout.HasChars
   build-depends:       base >= 4.12 && < 5,
                        text,
                        containers,