--- a/doclayout.cabal
+++ b/doclayout.cabal
@@ -24,7 +24,9 @@ cabal-version:       >=1.10
 library
   hs-source-dirs:      src
   exposed-modules:     Text.DocLayout
-  other-modules:       Text.DocLayout.HasChars
+  other-modules:       Text.DocLayout.HasChars,
+                       Text.DocLayout.ANSIFont,
+                       Text.DocLayout.Attributed
   build-depends:       base >= 4.12 && < 5,
                        text,
                        containers,