@jgm 2024-03-06
I'm running benchmarks:
after this PR:
```
benchmarking sample document 2
time                 11.13 μs   (11.11 μs .. 11.14 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 11.12 μs   (11.10 μs .. 11.15 μs)
std dev              67.15 ns   (35.39 ns .. 121.9 ns)

benchmarking reflow English
time                 279.5 μs   (277.6 μs .. 282.0 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 279.2 μs   (278.0 μs .. 281.9 μs)
std dev              5.800 μs   (3.359 μs .. 10.51 μs)
variance introduced by outliers: 14% (moderately inflated)

benchmarking reflow Greek
time                 377.6 μs   (373.6 μs .. 382.3 μs)
                     0.999 R²   (0.999 R² .. 0.999 R²)
mean                 377.1 μs   (374.5 μs .. 380.8 μs)
std dev              10.27 μs   (7.674 μs .. 16.40 μs)
variance introduced by outliers: 20% (moderately inflated)

benchmarking tabular English
time                 5.521 ms   (5.480 ms .. 5.574 ms)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 5.498 ms   (5.478 ms .. 5.521 ms)
std dev              67.53 μs   (52.46 μs .. 87.20 μs)

benchmarking tabular Greek
time                 7.715 ms   (7.104 ms .. 8.469 ms)
                     0.974 R²   (0.962 R² .. 0.999 R²)
mean                 7.154 ms   (7.035 ms .. 7.408 ms)
std dev              469.4 μs   (209.1 μs .. 792.7 μs)
variance introduced by outliers: 36% (moderately inflated)

benchmarking soft spaces at end of line
time                 5.711 μs   (5.676 μs .. 5.776 μs)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 5.742 μs   (5.717 μs .. 5.778 μs)
std dev              99.33 ns   (80.64 ns .. 140.8 ns)
variance introduced by outliers: 16% (moderately inflated)
```

vs before:
```
benchmarking sample document 2
time                 13.54 μs   (13.51 μs .. 13.59 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 13.56 μs   (13.52 μs .. 13.65 μs)
std dev              197.0 ns   (115.7 ns .. 285.9 ns)
variance introduced by outliers: 11% (moderately inflated)

benchmarking reflow English
time                 123.4 μs   (123.3 μs .. 123.5 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 123.3 μs   (123.2 μs .. 123.4 μs)
std dev              315.7 ns   (252.8 ns .. 379.8 ns)

benchmarking reflow Greek
time                 108.9 μs   (108.7 μs .. 109.1 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 109.6 μs   (109.4 μs .. 109.8 μs)
std dev              802.4 ns   (661.6 ns .. 980.5 ns)

benchmarking tabular English
time                 1.523 ms   (1.511 ms .. 1.542 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 1.527 ms   (1.524 ms .. 1.532 ms)
std dev              13.76 μs   (9.709 μs .. 24.32 μs)

benchmarking tabular Greek
time                 1.901 ms   (1.896 ms .. 1.907 ms)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 1.918 ms   (1.908 ms .. 1.936 ms)
std dev              44.25 μs   (26.99 μs .. 73.68 μs)
variance introduced by outliers: 11% (moderately inflated)

benchmarking soft spaces at end of line
time                 5.136 μs   (5.128 μs .. 5.146 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 5.133 μs   (5.127 μs .. 5.140 μs)
std dev              20.13 ns   (16.38 ns .. 27.61 ns)

```