@jgm 2024-03-06
The "reflow" and "tabular" benchmarks show a very significant slowdown, which might be reason to revert or rethink this change.  I'll see if I can quantify the impact on pandoc's benchmarks.