@silby 2024-03-06
I’m not altogether surprised it’s slow, considering that I’m wrapping what would otherwise be a bunch of variously-optimized stringlike operations in a recursive ADT, and then doing a whole additional pass over it to render it to the final result. Sorry I didn’t think to quantify that before sending the PR.

My design here is pretty much shaped by the fact that Blocks have to be prerendered to _something_ that is `HasChars`, measured with `realLength`, then chopped up into lines that all carry the font information with them. I didn’t want the block alignment to simply not work right in the ANSI context.

It seems like less-naïve implementation of `Attributed a` should be possible, that’s not just built out of raw cons cells that might recurse in either direction.

Maybe we can also win some performance back by replacing `renderANSI = attrRender . prerender` with a version that knows how to skip the `Attributed` step on request and just emit strings to the output. I think I’ll need another typeclass though.