@jgm 2024-03-06
OK, I can now quantify the impact on pandoc:  a few selected examples from the writer benchmarks.

asciidoctor: from 3.2ms to 4.19ms
commonmark: 5.3ms to 7.96 ms
djot: 2.76 ms to 6.3 ms
docbook5: 3.4ms to 14.1 ms
html: 3.64ms to 8.28ms
icml: 14.7ms to 90.1 ms
latex: 2.78 ms to 7.40 ms
man: 1.85 ms to 3.64 ms
markdown: 5.67 ms to 9.00 ms
mediawiki (which doesn't use doclayout): 1.41 ms to 1.45 ms
opendocument: 8.70ms to 27.7ms
org: 1.76ms to 5.37ms