@jgm 2024-03-06
Do you mean replace the definition of `renderPlain` ?