@jgm 2024-03-06
I ran the benchmarks with profiling enabled and got the following profiteur report:

<img width="753" alt="image" src="https://github.com/jgm/doclayout/assets/3044/0108afa6-ddf6-4c21-8c1d-47fc417bf751">