@silby 2024-03-06
[My branch `perf` over here](https://github.com/silby/doclayout/commits/perf/) replaces the cons-based `Attributed a` with a newtype around a `Seq` of standalone `Attr a` thingies. Not moving that to a PR tonight b/c I was just bashing it til it worked, and I have no idea if using `Seq` is even a useful optimization, and so forth, but it does improve performance: the `reflow English` and `reflow Greek` benchmarks are comparable to baseline instead of taking twice as long, and the `tabular English` and `tabular Greek` benchmarks are 1.7-2x instead of 3.5x over baseline. If you want to pull that down now and try yourself go ahead. More to come.