@silby 2024-03-06
looks like with the above-mentioned branch of `DocLayout` there are no performance regressions in the pandoc benchmarks vs baseline. I tried putting some extra tables in `testsuite.txt` just in case, still ok. From cursory inspection there seem to be no pandoc writers that exercise the `lblock`/`rblock`/`cblock` functions of `DocLayout` as creatively as the `tabular` benchmarks here do. Now I'm actually done for the night.