@jgm 2024-03-06
That's really great!  I'll wait for the PR.